LayoutTests/fast/reflections/reflection-computed-style-expected.txt

Test calling getPropertyValue on computed styles for -webkit-border-image property.

-webkit-box-reflect: none;
  -webkit-box-reflect
    getPropertyValue:    none
    getPropertyCSSValue: [object CSSPrimitiveValue]

-webkit-box-reflect: below;
  -webkit-box-reflect
    getPropertyValue:    below 0px none
    getPropertyCSSValue: [object CSSValue]

-webkit-box-reflect: right 50%;
  -webkit-box-reflect
    getPropertyValue:    right 50% none
    getPropertyCSSValue: [object CSSValue]

-webkit-box-reflect: below 5px -webkit-gradient(linear, 0 0, 0 0) 25 25 25 25 stretch stretch;
  -webkit-box-reflect
    getPropertyValue:    below 5px -webkit-gradient(linear, 0 0, 0 0) 25 fill / auto / 0px stretch
    getPropertyCSSValue: [object CSSValue]
